Pytorch/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
pycocotools>=2.0.6
matplotlib>=3.7.1
tqdm>=4.65.0
pillow>=9.5.0
numpy>=1.24.0